Rename aggregators folder
